+++ cca-manifest-logic/src/plugin-maps.js
@@ -10,7 +10,6 @@ var DEFAULT_PLUGINS = [
     'org.apache.cordova.inappbrowser',
     'org.apache.cordova.network-information',
     'org.apache.cordova.statusbar',
-    'org.apache.cordova.labs.keyboard',
     'org.chromium.common',
     'org.chromium.runtime',
     'org.chromium.storage',