+++ chrome-cordova/chrome-apps-api-tests/tests/auto/test.chrome.audioCapture.js
@@ -29,7 +29,7 @@ registerAutoTests("chrome.audioCapture", function() {
   }
 
   beforeEach(function(done) {
-    addMatchers(customMatchers);
+    jasmine.addMatchers(customMatchers);
     done();
   });