+++ chrome-cordova/chrome-apps-api-tests/tests/auto/test.chrome.system.cpu.js
@@ -41,7 +41,7 @@ registerAutoTests('chrome.system.cpu', function() {
   };
   
   beforeEach(function(done) {
-    addMatchers(customMatchers);
+    jasmine.addMatchers(customMatchers);
     done();
   });