+++ chrome-cordova/chrome-apps-api-tests/tests/auto/test.chrome.system.network.js
@@ -31,7 +31,7 @@ registerAutoTests('chrome.system.network', function() {
   };
   
   beforeEach(function(done) {
-    addMatchers(customMatchers);
+    jasmine.addMatchers(customMatchers);
     done();
   });