+++ chrome-cordova/plugins/chrome.sockets.tcp/src/android/ChromeSocketsTcp.java
@@ -1,6 +1,7 @@
 package org.chromium;
 
 import java.io.IOException;
+import java.io.OutputStream;
 import java.net.InetSocketAddress;
 import java.net.SocketException;
 import java.nio.ByteBuffer;
@@ -28,6 +29,7 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import android.net.Uri;
 import android.util.Log;
 
 public class ChromeSocketsTcp extends CordovaPlugin {
@@ -81,12 +83,16 @@ public class ChromeSocketsTcp extends CordovaPlugin {
     return true;
   }
 
-  public void onDestory() {
+  @Override
+  public void onDestroy() {
+    super.onDestroy();
     closeAllSockets();
     stopSelectorThread();
   }
 
+  @Override
   public void onReset() {
+    super.onReset();
     closeAllSockets();
     stopSelectorThread();
   }
@@ -102,9 +108,16 @@ public class ChromeSocketsTcp extends CordovaPlugin {
     return error;
   }
 
+  private void sendReceiveEvent(PluginResult result) {
+    if (recvContext != null) {
+      result.setKeepCallback(true);
+      recvContext.sendPluginResult(result);
+    }
+  }
+
   public int registerAcceptedSocketChannel(SocketChannel socketChannel)
       throws IOException {
-    TcpSocket socket = new TcpSocket(nextSocket++, recvContext, socketChannel);
+    TcpSocket socket = new TcpSocket(nextSocket++, socketChannel);
     sockets.put(Integer.valueOf(socket.getSocketId()), socket);
 
     addSelectorMessage(socket, SelectorMessageType.SO_ACCEPTED, null);
@@ -114,34 +127,47 @@ public class ChromeSocketsTcp extends CordovaPlugin {
 
   private void create(CordovaArgs args, final CallbackContext callbackContext)
       throws JSONException {
-    JSONObject properties = args.getJSONObject(0);
+    final JSONObject properties = args.getJSONObject(0);
 
-    try {
-      TcpSocket socket = new TcpSocket(nextSocket++, recvContext, properties);
-      sockets.put(Integer.valueOf(socket.getSocketId()), socket);
-      callbackContext.success(socket.getSocketId());
-    } catch (SocketException e) {
-    } catch (IOException e) {
-    }
+    final int socketId = nextSocket++;
+
+    // Use a background thread beacause TcpSocket constructor may perform IO operations
+    cordova.getThreadPool().execute(new Runnable() {
+        public void run() {
+          try {
+            TcpSocket socket = new TcpSocket(socketId, properties);
+            sockets.put(Integer.valueOf(socket.getSocketId()), socket);
+            callbackContext.success(socket.getSocketId());
+          } catch (SocketException e) {
+          } catch (IOException e) {
+          } catch (JSONException e) {
+          }
+        }
+      });
   }
 
   private void update(CordovaArgs args, final CallbackContext callbackContext)
       throws JSONException {
     int socketId = args.getInt(0);
-    JSONObject properties = args.getJSONObject(1);
-
-    TcpSocket socket = sockets.get(Integer.valueOf(socketId));
+    final JSONObject properties = args.getJSONObject(1);
+    final TcpSocket socket = sockets.get(Integer.valueOf(socketId));
 
     if (socket == null) {
       Log.e(LOG_TAG, "No socket with socketId " + socketId);
       return;
     }
 
-    try {
-      socket.setProperties(properties);
-      callbackContext.success();
-    } catch (SocketException e) {
-    }
+    // Use a background thread because setProperties may perform IO operations.
+    cordova.getThreadPool().execute(new Runnable() {
+        public void run() {
+          try {
+            socket.setProperties(properties);
+            callbackContext.success();
+          } catch (SocketException e) {
+          } catch (JSONException e) {
+          }
+        }
+      });
   }
 
   private void setPaused(CordovaArgs args, final CallbackContext callbackContext)
@@ -365,27 +391,17 @@ public class ChromeSocketsTcp extends CordovaPlugin {
   }
 
   private void startSelectorThread() {
-    if (selector != null && selectorThread != null) return;
-    try {
-      selector = Selector.open();
-      selectorThread = new SelectorThread(selector, selectorMessages, sockets);
-      selectorThread.start();
-    } catch (IOException e) {
-      selector = null;
-      selectorThread = null;
-      PluginResult err = new PluginResult(Status.ERROR, buildErrorInfo(-9, e.getMessage()));
-      err.setKeepCallback(true);
-      recvContext.sendPluginResult(err);
-    }
+    if (selectorThread != null) return;
+    selectorThread = new SelectorThread(selectorMessages, sockets);
+    selectorThread.start();
   }
 
   private void stopSelectorThread() {
-    if (selector == null && selectorThread == null) return;
+    if (selectorThread == null) return;
 
     addSelectorMessage(null, SelectorMessageType.T_STOP, null);
     try {
       selectorThread.join();
-      selector = null;
       selectorThread = null;
     } catch (InterruptedException e) {
     }
@@ -428,15 +444,13 @@ public class ChromeSocketsTcp extends CordovaPlugin {
   }
 
   private class SelectorThread extends Thread {
-    private final Selector selector;
     private BlockingQueue<SelectorMessage> selectorMessages;
     private Map<Integer, TcpSocket> sockets;
     private boolean running = true;
 
     SelectorThread(
-        Selector selector, BlockingQueue<SelectorMessage> selectorMessages,
+        BlockingQueue<SelectorMessage> selectorMessages,
         Map<Integer, TcpSocket> sockets) {
-      this.selector = selector;
       this.selectorMessages = selectorMessages;
       this.sockets = sockets;
     }
@@ -497,7 +511,19 @@ public class ChromeSocketsTcp extends CordovaPlugin {
     }
 
     public void run() {
+
+      try {
+        selector = Selector.open();
+      } catch (IOException e) {
+        throw new RuntimeException(e);
+      }
+
+      // process possible messages that send during openning the selector
+      // before select.
+      processPendingMessages();
+
       Iterator<SelectionKey> it;
+
       while (running) {
 
         try {
@@ -545,7 +571,6 @@ public class ChromeSocketsTcp extends CordovaPlugin {
 
   private class TcpSocket {
     private final int socketId;
-    private final CallbackContext recvContext;
 
     private SocketChannel channel;
 
@@ -566,14 +591,16 @@ public class ChromeSocketsTcp extends CordovaPlugin {
     private boolean persistent;
     private String name;
     private int bufferSize;
+    private Uri destUri;
+    private OutputStream destOutputStream;
+    private boolean append;
 
     private CallbackContext connectCallback;
     private CallbackContext secureCallback;
 
-    TcpSocket(int socketId, CallbackContext recvContext, JSONObject properties)
+    TcpSocket(int socketId, JSONObject properties)
         throws JSONException, IOException {
       this.socketId = socketId;
-      this.recvContext = recvContext;
 
       channel = SocketChannel.open();
       channel.configureBlocking(false);
@@ -587,10 +614,9 @@ public class ChromeSocketsTcp extends CordovaPlugin {
       setBufferSize();
     }
 
-    TcpSocket(int socketId, CallbackContext recvContext, SocketChannel acceptedSocket)
+    TcpSocket(int socketId, SocketChannel acceptedSocket)
         throws IOException {
       this.socketId = socketId;
-      this.recvContext = recvContext;
 
       channel = acceptedSocket;
       channel.configureBlocking(false);
@@ -603,11 +629,22 @@ public class ChromeSocketsTcp extends CordovaPlugin {
       paused = true;
     }
 
-    void setDefaultProperties() {
+    void resetDestOutputStream() throws IOException {
+      if (destOutputStream != null) {
+        destOutputStream.close();
+        destOutputStream = null;
+        destUri = null;
+      }
+    }
+
+    void setDefaultProperties() throws IOException {
       paused = false;
       persistent = false;
       bufferSize = 4096;
       name = "";
+      destUri = null;
+      append = false;
+      resetDestOutputStream();
     }
 
     // Only call this method on selector thread
@@ -647,6 +684,30 @@ public class ChromeSocketsTcp extends CordovaPlugin {
       if (!properties.isNull("bufferSize")) {
         bufferSize = properties.getInt("bufferSize");
         setBufferSize();
+      }
+
+      if (!properties.isNull("append"))
+        append = properties.getBoolean("append");
+
+      if (!properties.isNull("destUri")) {
+        try {
+          resetDestOutputStream();
+
+          String uriString = properties.getString("destUri");
+
+          if (uriString.length() > 0) {
+            Uri uri = Uri.parse(uriString);
+            destOutputStream = webView.getResourceApi().openOutputStream(uri, append);
+
+            // Only update the destUri if the input uri is valid for openOutputStream()
+            destUri = uri;
+          }
+
+        } catch (IOException e) {
+          JSONObject info = buildErrorInfo(-2, e.getMessage());
+          info.put("socketId", socketId);
+          sendReceiveEvent(new PluginResult(Status.ERROR, info));
+        }
       }
     }
 
@@ -703,6 +764,7 @@ public class ChromeSocketsTcp extends CordovaPlugin {
     void disconnect() throws IOException {
       if (key != null && channel.isRegistered())
         key.cancel();
+      resetDestOutputStream();
       channel.close();
     }
 
@@ -921,6 +983,11 @@ public class ChromeSocketsTcp extends CordovaPlugin {
       info.put("name", name);
       info.put("paused", paused);
 
+      if (destUri != null) {
+        info.put("append", append);
+        info.put("destUri", destUri.toString());
+      }
+
       if (channel.socket().getLocalAddress() != null) {
         info.put("localAddress", channel.socket().getLocalAddress().getHostAddress());
         info.put("localPort", channel.socket().getLocalPort());
@@ -946,14 +1013,14 @@ public class ChromeSocketsTcp extends CordovaPlugin {
 
       // This may cause starvation if multiple sockets are trying to reading
       // large amount of data at same time.
-      if (!isReadyToRead) {
+      if (!isReadyToRead && destOutputStream == null) {
         return bytesRead;
       }
 
       try {
         bytesRead = channel.read(receiveDataBuffer);
         if (bytesRead < 0)
-          return bytesRead;
+          throw new IOException("Socket closed by remote peer");
 
         if (sslEngine != null) {
 
@@ -972,31 +1039,32 @@ public class ChromeSocketsTcp extends CordovaPlugin {
       } catch (IOException e) {
         JSONObject info = buildErrorInfo(-2, e.getMessage());
         info.put("socketId", socketId);
-        PluginResult errResult = new PluginResult(Status.ERROR, info);
-        errResult.setKeepCallback(true);
-        recvContext.sendPluginResult(errResult);
+        sendReceiveEvent(new PluginResult(Status.ERROR, info));
       } catch (JSONException e) {
       }
       return bytesRead;
     }
 
-    private void sendReceive(ByteBuffer data) throws JSONException {
+    private void sendReceive(ByteBuffer data) throws JSONException, IOException {
 
       // TODO: avoid this copy by creating a new PluginResult overload.
       byte[] recvBytes = new byte[data.limit()];
       data.get(recvBytes);
 
-      PluginResult dataResult = new PluginResult(Status.OK, recvBytes);
-      dataResult.setKeepCallback(true);
-      recvContext.sendPluginResult(dataResult);
+      JSONObject info = new JSONObject();
+      info.put("socketId", socketId);
 
-      JSONObject metadata = new JSONObject();
-      metadata.put("socketId", socketId);
-      PluginResult metadataResult = new PluginResult(Status.OK, metadata);
-      metadataResult.setKeepCallback(true);
-
-      isReadyToRead = false;
-      recvContext.sendPluginResult(metadataResult);
+      if (destOutputStream != null) {
+        destOutputStream.write(recvBytes);
+        destOutputStream.flush();
+        info.put("destUri", destUri.toString());
+        info.put("bytesRead", recvBytes.length);
+        sendReceiveEvent(new PluginResult(Status.OK, info));
+      } else {
+        sendReceiveEvent(new PluginResult(Status.OK, info));
+        isReadyToRead = false;
+        sendReceiveEvent(new PluginResult(Status.OK, recvBytes));
+      }
     }
 
     private class TcpSendPacket {